+++ history.md
@@ -1,3 +1,6 @@
+* Widgets can now capture events, preventing other widgets from responding to same event
+* Fixed bug with combat interface scrolling
+* Mission will check environment, and warn PC if any lancemates cannot take part
 * Added Androbot and Forgebot monsters
 * Immobilized lancemates get left behind if their mecha cannot be repaired
 * When placing gears on map, the environment is taken into account